requirements.txt
requests>=2.31.0
beautifulsoup4>=4.12.2
readability-lxml>=0.8.1
lxml>=4.9.3
PyYAML>=6.0.1
pytest>=7.4.0
# Optional for JS rendering:
# playwright>=1.41.0
